Extracts markets into one market per contract instance.
This change makes working with the code much easier and gives us the distinct advantage of being able to leverage `REPUTATION.balanceOf(self)` to great effect (same with other tokens).

Also started on the rename/merging of events => markets.

Filled in a number of TODOs and FIXMEs and added some more as I noticed them.